pyproject.toml
[project]
name = "codeit-ai-1team-llm-project"
version = "0.1.0"
description = "코드잇 중급 프로젝트 RAG 문서 요약 챗봇"
authors = [
    {name = "dong1203", email = "hambur1203@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12.3"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12.3"
# LangChain 생태계
langchain = "^1.0.5"
langchain-core = "^1.0.4"
langchain-text-splitters = "^1.0.0"
langchain-openai = "^1.0.2"

# 문서 처리
pypdf = "^6.2.0"
olefile = "^0.47"

# 데이터 처리
pandas = "^2.3.3"
numpy = "^2.3.0"

# OpenAI API
openai = "^2.7.2"

# 벡터 DB
chromadb = "^1.0.0"

# 시각화
streamlit = "^1.40.0"
plotly = "^6.0.0"

# 차원 축소
scikit-learn = "^1.5.0"

# 유틸리티
tqdm = "^4.67.1"
python-dotenv = "^1.0.0"
langchain-chroma = "^1.0.0"
openpyxl = "^3.1.5"
wandb = "^0.23.0"
ragas = "^0.3.9"
langsmith = "^0.4.43"
datasets = "^4.4.1"
langchain-community = "^0.4.1"
rapidfuzz = "^3.14.3"
matplotlib = "^3.10.7"
rank-bm25 = "^0.2.2"
sentence-transformers = "^5.1.2"
flagembedding = "^1.3.5"
llama-cpp-python = "^0.3.16"
huggingface-hub = ">=0.20.0"

[tool.poetry.group.dev.dependencies]
# 개발 도구 (선택사항)
pytest = "^8.0.0"
black = "^24.0.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"